fix(uploader): demo拆解与规范
