+++ src/packages/uploader/demo.scss
@@ -1,8 +1,3 @@
 .demo-uploader {
   padding: 57px 17px 15px 17px !important;
 }
-.demo {
-  .cell-wrap {
-    flex-wrap: wrap;
-  }
-}
