+++ src/packages/uploader/demos/h5/demo13.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { Uploader } from '@nutui/nutui-react'
+
+const Demo13 = () => {
+  return <Uploader disabled />
+}
+export default Demo13
