+++ src/packages/uploader/demos/h5/demo5.tsx
@@ -1,0 +1,27 @@
+import React, { useState } from 'react'
+import { Uploader, Button, Progress } from '@nutui/nutui-react'
+
+const Demo5 = () => {
+  const uploadUrl = 'https://my-json-server.typicode.com/linrufeng/demo/posts'
+  const [progressPercent, setProgressPercent] = useState(0)
+  const onProgress = ({ event, options, percentage }: any) => {
+    setProgressPercent(percentage)
+  }
+
+  return (
+    <>
+      <Uploader url={uploadUrl} onProgress={onProgress}>
+        <Button type="success" size="small">
+          上传文件
+        </Button>
+      </Uploader>
+      <br />
+      <Progress
+        percentage={progressPercent}
+        strokeColor="linear-gradient(270deg, rgba(18,126,255,1) 0%,rgba(32,147,255,1) 32.815625%,rgba(13,242,204,1) 100%)"
+        status
+      />
+    </>
+  )
+}
+export default Demo5
