+++ src/packages/uploader/demos/h5/demo9.tsx
@@ -1,0 +1,45 @@
+import React from 'react'
+import { Uploader } from '@nutui/nutui-react'
+
+const Demo9 = () => {
+  const uploadUrl = 'https://my-json-server.typicode.com/linrufeng/demo/posts'
+  const canvastoFile = (
+    canvas: HTMLCanvasElement,
+    type: string,
+    quality: number
+  ): Promise<Blob | null> => {
+    return new Promise((resolve) => {
+      canvas.toBlob((blob) => resolve(blob), type, quality)
+    })
+  }
+  const fileToDataURL = (file: Blob): Promise<any> => {
+    return new Promise((resolve) => {
+      const reader = new FileReader()
+      reader.onloadend = (e) => resolve((e.target as FileReader).result)
+      reader.readAsDataURL(file)
+    })
+  }
+  const dataURLToImage = (dataURL: string): Promise<HTMLImageElement> => {
+    return new Promise((resolve) => {
+      const img = new Image()
+      img.onload = () => resolve(img)
+      img.src = dataURL
+    })
+  }
+  const beforeUpload = async (files: File[]) => {
+    const canvas = document.createElement('canvas')
+    const context = canvas.getContext('2d') as CanvasRenderingContext2D
+    const base64 = await fileToDataURL(files[0])
+    const img = await dataURLToImage(base64)
+    canvas.width = img.width
+    canvas.height = img.height
+    context.clearRect(0, 0, img.width, img.height)
+    context.drawImage(img, 0, 0, img.width, img.height)
+    const blob = (await canvastoFile(canvas, 'image/jpeg', 0.5)) as Blob
+    const f = await new File([blob], files[0].name, { type: files[0].type })
+    return [f]
+  }
+
+  return <Uploader url={uploadUrl} multiple beforeUpload={beforeUpload} />
+}
+export default Demo9
