+++ src/packages/uploader/demos/taro/demo13.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+import { Uploader } from '@nutui/nutui-react-taro'
+
+const Demo13 = () => {
+  return <Uploader disabled />
+}
+export default Demo13
